Fixes for Bugzilla bugs from testing ISO Assignment tool
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@11690 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/plugins/ipadexporter/iPadDBExporter.java
@@ -3804,8 +3804,11 @@ public class iPadDBExporter implements VerifyCollectionListener
         if (StringUtils.isNotEmpty(institutionImageName))
         {
             File   srcPicFile  = new File(UIRegistry.getAppDataDir() + File.separator + institutionImageName);
-            FileUtils.copyFileToDirectory(srcPicFile, cacheDir);
-            fileNamesForExport.add(new ChartFileInfo(institutionImageName, "Institution Picture", "en"));
+            if (srcPicFile.exists())
+            {
+                FileUtils.copyFileToDirectory(srcPicFile, cacheDir);
+                fileNamesForExport.add(new ChartFileInfo(institutionImageName, "Institution Picture", "en"));
+            }
         }
         
         dbUpdateConn = DBConnection.getInstance().createConnection();
